Assign identity to lateral horn neurons

Given a neuron, this function assigns it to a lateral horn primary neurite tract, anatomy group and cell type as described by Frechter et al. 2018. Note should be taken with what side of the brain the sample group is on. Note that the function will force an assignation even if the neurons in someneuronlist belong to unidentified tracts/anatomy group/cell types. Relies on NBLAST. Skeletons must have their somas as their root. This function provides a good guess, however if it gets a neurons assignment incorrect, that does not necessarily mean the neuron in question is not covered by the naming system.

assign_lh_neuron(someneuronlist, most.lhns = lhns::most.lhns,
  most.lhns.dps = catnat::most.lhns.dps,
  most.lhns.pnts.dps = catnat::most.lhns.pnts.dps, brain = NULL)

scan_lh_matches(someassignedneuronlist)

Arguments

someneuronlist
a neuronlist object

most.lhns
a dataset containing example neurons of different primary neurite tracts, anatomy groups and cell types. Defaults to the FlyCircuit neurons and dye-fills used in Frechter et al. 2018, as do the two arguments below

most.lhns.dps
a dotprops object of the above

most.lhns.pnts.dps
the primary neurite tracts to search against as a dotprops object

brain
the brainspace of the someneuronlist. If left NULL, assumes the space is FCWB

someassignedneuronlist
a neuronlist that has been passed through assign_lh_neuron

...
additional arguments passed to methods

Value

a neuronlist with the best guess for primary neurite tract, anatomy group and cell type listed in its metadata. Quality of this estimation depends on quality of the skeleton objects used in this function and their registration ot FCWB space