Try to detect whether a neuron your are about to import into CATMAID might cause a duplication in that CATMAID instance

Programmatically search for skeleton IDs pertaining to neurons within a search volume defined by a bounding box.

catmaid_duplicated(neuron, skid = 0, tolerance = NULL,
  duplication.range.nm = 20, downsample = 1, fafbseg = FALSE,
  pid = 1, conn = NULL, ...)

Arguments

neuron
the neuron object you are thinking about uploading

skid
the skeleton ID that corresponds to neuron

tolerance
how many potentially duplicated nodes you will tolerate. If NULL, skeleton IDs for potentially duplicated neurons are returned.

duplication.range.nm
determines the size of the bounding box around each node in neuron to search for a duplicated. Defaults to 10 nm

downsample
if downsample > 1, a sample of number.of.points/downsample is taken from the neuron, and only these points are used to assess duplication. Speeds things up

fafbseg
whether or not to use fafbseg::read_brainmaps_meshes on the TODO tag locations and restrict possible duplication to skeletons within the search range and within the cognate auto-segmented volume

pid
project id. Defaults to 1

conn
CATMAID connection object, see ?catmaid::catmaid_login for details

...
methods passed to catmaid::catmaid_fetch

Value

If tolerance is NULL, then a list, with entries duplicated.nodes (a TRUE or FALSE for ever node in neuron, TRUE is there is another skeleton within duplication.range.nm) and overlapping.skids, a list of skeleton IDs for potentially overlapping neurons. If tolerance is a numeric value between 0 and 1, TRUE or FALSE is returned.