Generate a connectivity matrix based on euclidean distance between points

Generates an 'overlap matrix' of overlap scores between neurons in the 'neurons' and 'targets' pools. For every point in a given neuron in 'neurons', a distance score is calculated to every point in a neuron in 'targets'. The sum of this score is added to the final output matrix. The score is calculated as e(-d^2/2δ^2), where d is the euclidean distance between the two points, and δ is the expected distance in um that is considered 'close'.

overlap.connectivity.matrix(neurons, targets, neuropil = NULL,
  delta = 1)

overlap.connectivity.matrix.catmaid(neurons, targets, neuropil = NULL,
  delta = 1, split = FALSE, rval = c("score.matrix", "neuronlist"))

Arguments

neurons
first set of neurons

targets
second set of neurons

neuropil
an as3d object of the neuropil in which to consider connectivity. Defaults to whole brain.

delta
the distance (in um) at which a synapse might occur

split
with a CATMAID neuron, whether or not to split the neuron using flow centrality

rval
whether to return an overlap score matrix or a neuronlist in which every point is given an overlap score against each of the target neurons

...
additional arguments passed to methods

Value

a matrix of 3D points