Plot neurons split up by flow centrality

Plot neurons split up by flow centrality

seesplit3d(someneuronlist, col = c("blue", "orange", "purple", "green",
  "grey", "pink"), splitnode = FALSE, WithConnectors = TRUE,
  WithNodes = F, soma = 100, highflow = F, lwd = 1, radius = 1,
  ...)

splitscan(someneuronlist, col = c("blue", "orange", "purple", "green",
  "grey", "pink"), WithConnectors = T, WithNodes = F, soma = 100,
  highflow = F, Verbose = T, Wait = T, sleep = 0.1,
  extrafun = NULL, selected_file = NULL, selected_col = "black",
  yaml = TRUE, ...)

Arguments

someneuronlist
a neuronlist or neuron object that has been modified by flow.centrality

col
colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.

WithConnectors
whether ot plot the anatomical location of pre (red) and post (cyan) synapses.

WithNodes
whether to plot branch points

soma
whether to plot a soma, and what the radius should be

highflow
wheather to plot the nodes of highest (with in one standard deviation less than maximum) flow centrality (pink points)

lwd
Line width (default 1)

radius
For connectors and axon-dendrite split node (default 1)

...
additional arguments passed to methods.

Verbose
logical indicating that info about each selected neuron should be printed (default TRUE)

Wait
logical indicating that there should be a pause between each displayed neuron

sleep
time to pause between each displayed neuron when Wait=TRUE

extrafun
an optional function called when each neuron is plotted, with two arguments: the current neuron name and the current selected neurons

selected_file
an optional path to a yaml file that already contains a selection

selected_col
the color in which selected neurons (such as those specified in selected_file) should be plotted

yaml
a logical indicating that selections should be saved to disk in (human-readable) yaml rather than (machine-readable) rda format

Value

Plots coloured neuron(s)

See also