Get data and plots describing tracer activity statistics

Some functions that quickly visualise tracer effort, or retrieve data to do with tracer effort.

tracer.plot(names = c("Alex Bates", "Ruairi Roberts", "Greg Jefferis",
  "Clement Hallou", "Philipp Ranft", "Philipp Schlegel", "Fiona Love",
  "Amelia Edmondson-Stait"), calc.date = "2016-04-01",
  from.date = "2016-04-01", to.date = Sys.Date(), cumulative = T,
  ...)

tracer.treenodes.plot(skids = NULL, names = c("Alex Bates",
  "Ruairi Roberts", "Greg Jefferis", "Clement Hallou", "Philipp Ranft",
  "Philipp Schlegel", "Fiona Love", "Amelia Edmondson-Stait"),
  from.date = "2016-04-01", to.date = Sys.Date(), ...)

tracer.neuron.stats(skids, value = c("nodes", "pre", "post"), ...)

Arguments

names
Full names of tracers

calc.date
The date from which to retrieve data for the graphs

from.date
The date from which to plot the graphs

to.date
The date to which to plot the graphs

cumulative
Whether to generate a cumulative plot. Defaults to true.

...
additional arguments passed to methods.

skids
The skids of the neurons to consider

value
Whether to look at tree node, pre synapse connector or post synaptic node data.

Value

tracer.plot() returns a plot for cable length traced and connectors placed by named tracers. tracer.treenodes.plot() visualises data on node generation.